move shared directory up a level
