+++ imports/blocks/give/fieldsets/shared/Icon.js
@@ -1,5 +1,5 @@
 import { PropTypes } from "react";
-import AccountType from "../../../../../components/accountType";
+import AccountType from "../../../../components/accountType";
 
 const Icon = ({ cardType }) => (
   <AccountType width="30px" height="21px" type={cardType} />
